Create Basic HuggingFacePar.py
To make the code multithreaded and parallelized, you can use the torch.nn.DataParallel module in PyTorch, which allows you to parallelize the model across multiple GPUs. Here's an example of how you can modify the above code to use DataParallel.
In this code, the nn.DataParallel module is used to parallelize the model across multiple GPUs. The to method is used to move the model to the specified device (either CPU or GPU). The input IDs are also moved to the device using the to method. The with torch.no_grad() context is used to disable gradient tracking during inference, which can speed up the generation process. Finally, the generated output is moved back to the CPU using the cpu() method and decoded using the tokenizer.